+++ node/src/common/middleware/request-id.middleware.ts
@@ -1,0 +1,17 @@
+import { Injectable, NestMiddleware } from '@nestjs/common';
+import { Request, Response, NextFunction } from 'express';
+import { v4 as uuidv4 } from 'uuid';
+
+@Injectable()
+export class RequestIdMiddleware implements NestMiddleware {
+  use(req: Request, res: Response, next: NextFunction) {
+    const requestId = uuidv4();
+
+    req.headers['x-request-id'] = requestId;
+    res.setHeader('X-Request-ID', requestId);
+
+    req.requestId = requestId;
+
+    next();
+  }
+}