+++ node/src/types/express.d.ts
@@ -3,5 +3,6 @@ import { CurrentUserDto } from '../user/dto/current-user.dto';
 declare module 'express' {
   interface Request {
     currentUser?: CurrentUserDto;
+    requestId?: string;
   }
 }